Complete initial mobile auth
