+++ client/mobile/components/documents/document.js
