+++ client/mobile/components/documents/sample_native_document.js
