+++ client/mobile/components/splash/auth_container.js
